+++ FBXWRapperDll/SceneContainer/FBXSCeneContainer.cpp
@@ -13,7 +13,7 @@ void wrapdll::FBXSCeneContainer::GetVertices(int meshindex, PackedCommonVertex**
     *ppVertices = m_packedMeshes[meshindex].vertices.data();
 }
 
-void wrapdll::FBXSCeneContainer::GetIndices(int meshindex, uint16_t** ppIndices, int* itemCount)
+void wrapdll::FBXSCeneContainer::GetIndices(int meshindex, uint32_t** ppIndices, int* itemCount)
 {
     *itemCount = static_cast<int>(m_packedMeshes[meshindex].indices.size());
     *ppIndices = m_packedMeshes[meshindex].indices.data();
